[EAK-396] Moved the XmlFactory class and lowered visibility
